\name{dnase2tf-package}
\alias{dnase2tf-package}
\docType{package}
\title{
finds footprint candidates from the DNase-I seq data.
}
\description{
dnase2tf finds footprint candidates from the DNase-I seq data (given by
datafilepath) on the user specified regions of interest (given by
hotspotfilename).  Mappability file location (mapfiledir) and the
filename for the output file are also required.
}
\details{
\tabular{ll}{
Package: \tab dnase2tf\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-09-19\cr
License: \tab GNU General Public License, v 3.0\cr
}
}
\author{
Songjoon Baek

Maintainer: Songjoon Baek <sojbaek@users.sourceforge.net>

}
\references{
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
